requirements.txt
numpy==1.19.5
torch==1.11.0
transformers==4.12.5
pytorch-crf==0.7.2
scikit-learn == 0.23.2